refactor: Reorganize pages structure and remove redundant routes directory
