+++ src/components/Panel/UserProfile/components/DeleteAccount/DeleteAccount.component.tsx
