+++ src/components/Panel/components/CourseCard/CourseCard.component.tsx
