+++ src/components/Panel/components/CoursesList/CoursesList.component.tsx
