+++ src/components/Panel/hooks/useCourses.hook.ts
