+++ src/components/Panel/types.ts
