+++ src/pages/index.ts
@@ -1,5 +1,5 @@
 export { default as HomePage } from './HomePage/HomePage';
 export { default as LearningPage } from './LearningPage/LearningPage';
-export { default as AccountPage } from './AccountPage/AccountPage';
+export { default as PanelPage } from './PanelPage/PanelPage';
 export { default as AuthPage } from './AuthPage/AuthPage';
 export { default as PaymentPage } from './PaymentPage/PaymentPage';
